requirements.txt
# ShadowHawk Platform - Core Dependencies
# Copyright (c) 2024 ShadowHawk Platform
# Licensed under the MIT License

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# CORS and middleware
python-dotenv==1.0.0
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9

# PDF Generation
reportlab==4.0.7
pypdf==3.17.1
weasyprint==60.1

# AI/LLM Integration
openai==1.3.7
anthropic==0.7.7
langchain==0.0.350
tiktoken==0.5.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# MITRE ATT&CK
mitreattack-python==3.0.3
stix2==3.0.1

# Sandboxing & Docker
docker==7.0.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1

# Logging
structlog==23.2.0

# Validation
email-validator==2.1.0
validators==0.22.0

# Caching
redis==5.0.1

# Monitoring
prometheus-client==0.19.0
